requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
email-validator==2.1.0.post1
APScheduler==3.10.1
python-docx==1.1.0
requests
gunicorn